customersuportvoice/pyproject.toml
[project]
name = "customersuportvoice"
version = "0.1.0"
description = "Voice-based assistant"
requires-python = ">=3.9,<3.12"

dependencies = [
  "streamlit",
  "sounddevice",
  "scipy",
  "fpdf",
  "openai",
  "whisper",
  "TTS==0.22.0",
  "python-dotenv"
]
